리뷰 데이터 저장 및 추천 시스템 트리거 로직 변경

+++ src/main/java/com/cloudingYo/barrierFree/place/entity/Place.java
@@ -1,10 +1,15 @@
 package com.cloudingYo.barrierFree.place.entity;
 
-import jakarta.persistence.*;
-import jakarta.validation.constraints.NotNull;
-import lombok.*;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
-import static jakarta.persistence.GenerationType.*;
+import static jakarta.persistence.GenerationType.IDENTITY;
 
 @Entity
 @NoArgsConstructor
@@ -18,7 +23,7 @@ public class Place {
     @Column(name = "place_id")
     private Long id;
 
-    private int placeKey;
+    private Long placeKey;
     private String placename;
     private String en_placename;
     private String category;
@@ -29,6 +34,7 @@ public class Place {
      * 0: int값이라 별도의 설정이 없다면 기본값이 0
      */
     private int reviewCount;
+
     private int totalScroe;
 
     public Place(String placename) {
